Added 2 gems after some initial testing.
will_paginate to add basic pagination to the messages index as trying to render 500,000+ messages was a bit much and was killing my dev environment.
rink to add the ability for links in each message to be turned into clickable links when rendered. Really nice addition

Fixed an issue when saving the log file to enforce it being saved as a binary. There was a funky issue when trying to upload the full log file and this appears to solve it after doing some searching.