Added in the most basic of stats. Messages per hour.
Had to tweak the model as I couldn't figure out a way to get records by just an hour period of time while ignoring the date. Added both hour and minute columns which are populated when the record is created.

Also threw this into a partial to share across pages. Should do this with all stats views to make mixing them up easier.